requirements.txt
grpcio>=1.60.0
grpcio-tools>=1.60.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
